feat: add Tiingo data service with fallbacks and tests
